@@ -1,5 +1,5 @@
 module github.com/creachadair/taskgroup
 
-go 1.21
+go 1.22
 
 require github.com/fortytw2/leaktest v1.3.0
